Address payload validation review comments
