+++ tests/main/conftest.py
@@ -9,6 +9,7 @@ import sys
 import time
 from argparse import Namespace
 from collections.abc import Callable, Generator, Sequence
+from contextlib import contextmanager
 from pathlib import Path
 from typing import Any, Literal
 
@@ -669,6 +670,39 @@ def _import_package(output_path: Path) -> None:
             sys.modules.pop(module_name, None)
 
 
+@contextmanager
+def _generated_model(output_path: Path, module_name: str, model_name: str) -> Generator[Any, None, None]:
+    spec = importlib.util.spec_from_file_location(module_name, output_path)
+    if spec is None or spec.loader is None:  # pragma: no cover
+        pytest.fail(f"Unable to load generated module from {output_path}", pytrace=False)
+
+    module = importlib.util.module_from_spec(spec)
+    previous_module = sys.modules.get(spec.name)
+    sys.modules[spec.name] = module
+    try:
+        spec.loader.exec_module(module)
+        yield getattr(module, model_name)
+    finally:
+        if previous_module is None:
+            sys.modules.pop(spec.name, None)
+        else:
+            sys.modules[spec.name] = previous_module
+
+
+def _assert_model_json_invalid(model: Any, invalid_json: str, expected_error_type: str) -> None:
+    with pytest.raises(ValidationError) as exc_info:
+        model.model_validate_json(invalid_json)
+    errors = exc_info.value.errors()
+    if not errors:  # pragma: no cover
+        pytest.fail("Expected validation error but got an empty errors list", pytrace=False)
+    actual_error_type = errors[0]["type"]
+    if actual_error_type != expected_error_type:  # pragma: no cover
+        pytest.fail(
+            f"Expected validation error {expected_error_type!r}, got {actual_error_type!r}",
+            pytrace=False,
+        )
+
+
 def assert_generated_model_json_validation(
     output_path: Path,
     *,
@@ -681,16 +715,7 @@ def assert_generated_model_json_validation(
     expected_attribute_value: Any = None,
 ) -> None:
     """Import a generated module and validate JSON data through a generated Pydantic model."""
-    spec = importlib.util.spec_from_file_location(module_name, output_path)
-    if spec is None or spec.loader is None:  # pragma: no cover
-        pytest.fail(f"Unable to load generated module from {output_path}", pytrace=False)
-
-    module = importlib.util.module_from_spec(spec)
-    previous_module = sys.modules.get(spec.name)
-    sys.modules[spec.name] = module
-    try:
-        spec.loader.exec_module(module)
-        model = getattr(module, model_name)
+    with _generated_model(output_path, module_name, model_name) as model:
         parsed = model.model_validate_json(valid_json)
 
         if expected_attribute_path:
@@ -703,22 +728,7 @@ def assert_generated_model_json_validation(
                     pytrace=False,
                 )
 
-        with pytest.raises(ValidationError) as exc_info:
-            model.model_validate_json(invalid_json)
-        errors = exc_info.value.errors()
-        if not errors:  # pragma: no cover
-            pytest.fail("Expected validation error but got an empty errors list", pytrace=False)
-        actual_error_type = errors[0]["type"]
-        if actual_error_type != expected_error_type:  # pragma: no cover
-            pytest.fail(
-                f"Expected validation error {expected_error_type!r}, got {actual_error_type!r}",
-                pytrace=False,
-            )
-    finally:
-        if previous_module is None:
-            sys.modules.pop(spec.name, None)
-        else:
-            sys.modules[spec.name] = previous_module
+        _assert_model_json_invalid(model, invalid_json, expected_error_type)
 
 
 def assert_generated_model_json_invalid(
@@ -730,32 +740,8 @@ def assert_generated_model_json_invalid(
     expected_error_type: str,
 ) -> None:
     """Import a generated module and assert JSON data is rejected by a generated Pydantic model."""
-    spec = importlib.util.spec_from_file_location(module_name, output_path)
-    if spec is None or spec.loader is None:  # pragma: no cover
-        pytest.fail(f"Unable to load generated module from {output_path}", pytrace=False)
-
-    module = importlib.util.module_from_spec(spec)
-    previous_module = sys.modules.get(spec.name)
-    sys.modules[spec.name] = module
-    try:
-        spec.loader.exec_module(module)
-        model = getattr(module, model_name)
-        with pytest.raises(ValidationError) as exc_info:
-            model.model_validate_json(invalid_json)
-        errors = exc_info.value.errors()
-        if not errors:  # pragma: no cover
-            pytest.fail("Expected validation error but got an empty errors list", pytrace=False)
-        actual_error_type = errors[0]["type"]
-        if actual_error_type != expected_error_type:  # pragma: no cover
-            pytest.fail(
-                f"Expected validation error {expected_error_type!r}, got {actual_error_type!r}",
-                pytrace=False,
-            )
-    finally:
-        if previous_module is None:
-            sys.modules.pop(spec.name, None)
-        else:
-            sys.modules[spec.name] = previous_module
+    with _generated_model(output_path, module_name, model_name) as model:
+        _assert_model_json_invalid(model, invalid_json, expected_error_type)
 
 
 def run_main_url_and_assert(
