+++ tests/main/payload_validation/__init__.py
@@ -1,0 +1,18 @@
+"""Public helpers for schema-derived payload validation tests."""
+
+from __future__ import annotations
+
+from .cases import SCHEMA_CASES, discover_unaccounted_files
+from .codegen import load_generated_payload_adapter
+from .models import GeneratedModelCache, SchemaCase
+from .strategy import payload_strategy, validate_with_source_schema
+
+__all__ = [
+    "SCHEMA_CASES",
+    "GeneratedModelCache",
+    "SchemaCase",
+    "discover_unaccounted_files",
+    "load_generated_payload_adapter",
+    "payload_strategy",
+    "validate_with_source_schema",
+]
