+++ tests/main/payload_validation/constants.py
@@ -1,0 +1,147 @@
+"""Constants for schema-derived payload validation tests."""
+
+from __future__ import annotations
+
+from datamodel_code_generator.format import PythonVersion, is_supported_in_black
+from tests.main.conftest import CURRENT_PYTHON_VERSION
+
+_PAYLOAD_CLASS_NAME = "Payload"
+_CURRENT_PYTHON_VERSION = PythonVersion(CURRENT_PYTHON_VERSION)
+_SCHEMA_FILE_SUFFIXES = {".json", ".yaml", ".yml"}
+_COMBINED_SCHEMA_KEYS = ("allOf", "anyOf", "oneOf")
+_JSON_SCHEMA_KEYS = {
+    "$defs",
+    "$id",
+    "$ref",
+    "$schema",
+    "additionalProperties",
+    *_COMBINED_SCHEMA_KEYS,
+    "definitions",
+    "enum",
+    "items",
+    "properties",
+    "type",
+}
+_OBJECT_SCHEMA_KEYWORDS = {
+    "additionalProperties",
+    "maxProperties",
+    "minProperties",
+    "properties",
+    "propertyNames",
+    "required",
+    "unevaluatedProperties",
+}
+_ARRAY_SCHEMA_KEYWORDS = {
+    "additionalItems",
+    "contains",
+    "items",
+    "maxContains",
+    "maxItems",
+    "minContains",
+    "minItems",
+    "prefixItems",
+    "unevaluatedItems",
+    "uniqueItems",
+}
+_SIMPLE_ALLOF_IGNORED_KEYS = {"description", "title"}
+_SIMPLE_ALLOF_SUPPORTED_KEYS = {
+    "const",
+    "enum",
+    "exclusiveMaximum",
+    "exclusiveMinimum",
+    "format",
+    "maximum",
+    "maxItems",
+    "maxLength",
+    "minimum",
+    "minItems",
+    "minLength",
+    "multipleOf",
+    "type",
+    "uniqueItems",
+}
+_PAYLOAD_FORMAT_ENUMS = {
+    "byte": ["", "MA==", "Zm9v"],
+    "date-time-local": ["2023-12-25T10:30:00"],
+    "duration": ["P1D", "PT1S"],
+    "ipv4": ["127.0.0.1"],
+    "ipv6": ["::1"],
+    "uuid": ["00000000-0000-4000-8000-000000000000"],
+}
+_EXCLUDED_FILES: dict[str, str] = {
+    "jsonschema/allof_class_hierarchy.json": "intentionally invalid JSON fixture",
+    "jsonschema/non_dict_files/list_only.json": "input is JSON data, not a JSON Schema document",
+    "jsonschema/non_dict_files/list_only.yaml": "input is YAML data, not a JSON Schema document",
+    "jsonschema/non_dict_files/whitespace_only.yaml": "empty YAML fixture",
+    "jsonschema/non_json_object.json": "input is JSON data, not a JSON Schema document",
+    "jsonschema/null.json": "intentionally invalid JSON fixture",
+    "jsonschema/ref_to_json_list/list.json": "referenced JSON data list, not a schema document",
+    "jsonschema/unknown_format.json": "unknown format fixture intentionally emits a generator warning",
+    "openapi/complex_reference.json": "intentionally invalid JSON fixture",
+    "openapi/list.json": "input is JSON data, not an OpenAPI document",
+    "openapi/not.json": "intentionally invalid JSON fixture",
+    "openapi/subclass_enum.json": "intentionally invalid JSON fixture",
+}
+_EXCLUDED_CASES: dict[str, str] = {
+    "jsonschema/all_of_any_of_base_class_ref.json": "hypothesis-jsonschema cannot satisfy the allOf/anyOf constraints",
+    "openapi/allof.yaml::components.schemas.AllOfNested3": (
+        "hypothesis-jsonschema cannot satisfy the nested allOf component constraints"
+    ),
+    "openapi/allof_with_required_inherited_complex_allof.yaml::components.schemas.ProjectedItem": (
+        "hypothesis-jsonschema generates this nested allOf schema inconsistently"
+    ),
+    "openapi/allof_with_required_inherited_comprehensive.yaml::components.schemas.Entity": (
+        "hypothesis-jsonschema generates this inherited allOf schema inconsistently"
+    ),
+    "openapi/allof_with_required_inherited_comprehensive.yaml::components.schemas.ProjectedEntity": (
+        "hypothesis-jsonschema generates this inherited allOf schema inconsistently"
+    ),
+    "openapi/allof_with_required_inherited_coverage.yaml::components.schemas.EdgeCasesCoverage": (
+        "hypothesis-jsonschema generates this inherited allOf coverage schema inconsistently"
+    ),
+    "openapi/allof_with_required_inherited_coverage.yaml::components.schemas.EnumInAllOf": (
+        "primitive enum allOf fields need a compatibility-sensitive generator fix; "
+        "see payload_validation_future_work.md"
+    ),
+    "openapi/allof_with_required_inherited_coverage.yaml::components.schemas.MultipleOfBase": (
+        "primitive multipleOf allOf fields need a compatibility-sensitive generator fix; "
+        "see payload_validation_future_work.md"
+    ),
+    "openapi/allof_with_required_inherited_edge_cases.yaml::components.schemas.EdgeCases": (
+        "hypothesis-jsonschema generates this inherited allOf edge-case schema inconsistently"
+    ),
+    "openapi/allof_with_required_inherited_edge_cases.yaml::components.schemas.ProjectedEdgeCases": (
+        "hypothesis-jsonschema generates this allOf edge-case schema inconsistently"
+    ),
+    "openapi/discriminator.yaml::components.schemas.Demo": (
+        "discriminator branches overlap under JSON Schema validation; see payload_validation_future_work.md"
+    ),
+    "openapi/discriminator_float_mapping.yaml::components.schemas.Base": (
+        "numeric discriminator mapping needs an explicit backend compatibility policy; "
+        "see payload_validation_future_work.md"
+    ),
+    "openapi/discriminator_in_array_oneof.yaml::components.schemas.Demo": (
+        "array item discriminator tags conflict with inherited JSON Schema constraints; "
+        "see payload_validation_future_work.md"
+    ),
+    "openapi/discriminator_without_mapping.yaml::components.schemas.Demo": (
+        "implicit discriminator tags conflict with source enum constraints; see payload_validation_future_work.md"
+    ),
+    "openapi/ref_nullable.yaml::components.schemas.NullableChild": (
+        "top-level nullable object components need a wrapper policy; nullable refs are covered via Parent"
+    ),
+}
+
+
+def _payload_target_python_version() -> str:
+    if is_supported_in_black(_CURRENT_PYTHON_VERSION):
+        return CURRENT_PYTHON_VERSION
+    supported_versions = [
+        python_version
+        for python_version in PythonVersion
+        if python_version.version_key <= _CURRENT_PYTHON_VERSION.version_key and is_supported_in_black(python_version)
+    ]
+    return max(supported_versions, key=lambda python_version: python_version.version_key).value
+
+
+_PAYLOAD_TARGET_PYTHON_VERSION = _payload_target_python_version()
